feat: gmail templated
